feat: moved matrix features, fixed tests
