+++ src/renderer/features/matrix/MatrixLogin/index.ts
@@ -1,0 +1,1 @@
+export { MatrixLogin } from './ui/MatrixLogin';
