+++ src/renderer/features/matrix/MatrixVerification/index.ts
@@ -1,0 +1,1 @@
+export { MatrixVerification } from './ui/MatrixVerification';
