--- src/renderer/widgets/MatrixModal/index.ts
@@ -1,2 +1,0 @@
-export { MatrixLoginModal } from './ui/MatrixLoginModal';
-export { MatrixVerificationModal } from './ui/MatrixVerificationModal';
